fix: set the correct prefix
When inputs have an Upload types deeply nested, the uploaded
files where simple ignored and not forwarded to the micro-services.
This was happening, since the prefix variable on the extract function
was wrong. In order to correcly access the Upload the key variable
should be provided to the extract function as the prefix.