feat: add testing env
This commit the support for testing the file upload support in action.
The main object is to create a graphql API which accepts file as input
and returns the file content as a response. This is preferred over jest
testing, since we can't fully test connection between apollo gateway
and the micro service using jest.