feat: upgrade @apollo/gateway 2.0.4
Since @apollo/gateway 2.0.4 we cannot use plain `Request` instance
anymore, it's referred directly from the `node-fetch`, which is now a
peer dependency.

It should be harmless since the default `didEncounterError()`, the
only thing that really uses that, just ignores the parameter. Also,
the `node-fetch` version is a superset of the basic `Request`, so
things should work.